chore: 패키지구조 변경
Related: #70
