--- src/main/java/com/modoospace/config/auth/dto/AuthProviderConvertor.java
@@ -1,9 +1,0 @@
-package com.modoospace.config.auth.dto;
-
-import java.util.Map;
-
-@FunctionalInterface
-public interface AuthProviderConvertor {
-
-  OAuthAttributes convert(String userNameAttributeName, Map<String, Object> attributes);
-}
