+++ src/main/java/com/modoospace/config/auth/provider/AuthProviderConvertor.java
@@ -1,0 +1,10 @@
+package com.modoospace.config.auth.provider;
+
+import com.modoospace.config.auth.dto.OAuthAttributes;
+import java.util.Map;
+
+@FunctionalInterface
+public interface AuthProviderConvertor {
+
+    OAuthAttributes convert(String userNameAttributeName, Map<String, Object> attributes);
+}
